IT Job Search/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiosqlite==0.19.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
telethon==1.33.1
aiosmtplib==3.0.1
python-dotenv==1.0.0
cryptography==41.0.7